requirements.txt
openai>=1.30.0
firebase-admin>=6.5.0
telebot>=0.0.5
pydantic>=2.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2024.1
